Reinterpret flow direction at a higher level
Flow direction, as provided by the agent, relates to the underlying
interface, ie. the flow can be said as "ingress" when going from pod IP
to host interface, although at a higher level of abstraction it should
be considered egress.

This commit reinterpret flow directions based on source and host IPs and
reporter IP.

The initial flow direction is preserved, copied into a new field such as
IfDirection (for interface direction)