NETOBSERV-1275: Introduce new "INNER" direction for inner-node traffic (#483)
* Introduce new "INNER" direction for inner-node traffic

The flows (and duplicates) generated for inner-node traffic differs
compared to node-to-node traffic, and reinterpret direction isn't able
to decide between ingress or egress. This is causing discrepancies with
the dedup mechanism that filters out flows where Duplicate=true and also
favors ingress over egress.

To fix that, the proposed solution is to create this new INNER direction
specifically for this kind of traffic. Deduping this INNER traffic can
then rely solely on the Duplicate flag, since that flag was set from a
single Agent (single node) there will always be only one
Duplicate=false.

* update doc

* Enable reinterpret on conversations