NETOBSERV-1935: enable metrics from list/nested fields (#739)
* NETOBSERV-1935: enable metrics from list/nested fields

This makes it possible to generate metrics with labels/filters set on
list fields (e.g. interfaces) and nested fields (e.g. soon-coming
structured network events)

In metrics API, user needs to configure explicitly which list field needs to
be "flattened", in order to be consumable as filters/labels.

Nested fields can be consumed as filters/labels with the ">" character;

E.g: `flatten: [networkEvents], filters: [{key: "networkEvents>type", value: "acl"}], labels: [networkEvents>name]`
This is a sample config to filter a metric for ACL events and label it by name

* linter