Remove old contributors from OWNERS file, add new ones
We are refreshing the list of maintainers to keep it aligned with
the current state of contributions, removing accounts who haven't contributed in
the past year.

Note that the OWNERS file controls the interactions with prow, but does not affect individual permissions configured in github.
Flowlogs-pipeline was originally created by IBM Research (@eranra @KalmanMeth and @ronensc), even though it is today mostly developed by Red Hat.
This PR does not remove the permissions that they have through the
github config, which we can keep unchanged.